Lock driver while pools are imported.
zfsinstaller.exe uninstall would crash if unloaded while pools are still
imported.

+++ ZFSin/zfs/include/sys/zfs_ioctl.h
@@ -659,7 +659,7 @@ extern minor_t zfsdev_minor_alloc(void);
 
 extern int zfs_ioctl_osx_init(void);
 extern int zfs_ioctl_osx_fini(void);
-extern int zfs_ioc_unregister_fs(void);
+extern uint64_t zfs_ioc_unregister_fs(void);
 extern int zfs_vnop_force_formd_normalized_output;
 
 #endif	/* _KERNEL */
